Impl `Send` for `TryMutexGuard` when `T: Send`
The std-lib `Mutex` requires that the lock is released on the same thread it was acquired given interaction with the OS scheduler. Given that this implementation has no such requirement, the `MutexGuard` can be `Send` iff `T: Send`. `T` need-not be `Sync` as a `MutexGuard` represents an exclusive reference to the underlying item.